+++ all-is-cubes/tests/alloc.rs
@@ -2,11 +2,13 @@
 //!
 //! In a separate test crate to avoid modifying the global allocator elsewhere.
 
-use all_is_cubes::inv;
-use all_is_cubes::op::Operation;
 use allocation_counter::{AllocationInfo, measure};
 
 use all_is_cubes::block::{self, BlockAttributes};
+use all_is_cubes::inv;
+use all_is_cubes::math::{GridAab, GridPoint};
+use all_is_cubes::op::Operation;
+use all_is_cubes::space::Space;
 use all_is_cubes::universe::Universe;
 
 #[test]
@@ -55,7 +57,26 @@ fn clone_evaluated_block() {
     });
 }
 
-// TODO: Test cloning of `Operation`
+#[test]
+fn space_alloc_failure() {
+    // Ideally we would do this test with a custom allocator that is guaranteed to fail, but
+    // we don't have one of those handy.
+    // Instead, compute a size that will definitely fill all available memory, but fits in i32
+    // on the individual axes.
+    let size = (usize::MAX as f64).powf(3.0f64.recip()).floor() as u32;
+    assert!(
+        Space::builder(GridAab::from_lower_size(
+            GridPoint::splat(i32::MIN),
+            [size, size, size]
+        ))
+        .try_build()
+        .is_err()
+    );
+}
+
+// TODO: Test cloning of `Operation` for being zero-alloc
+
+// -------------------------------------------------------------------------------------------------
 
 #[track_caller]
 fn assert_no_alloc(f: impl FnOnce()) {